Add files via upload
Hi! My name is Naomi. I am a student at the university of North Carolina at charlotte and I am majoring in computer science. I am taking a course in software engineering this semester and I have a product that I will be talking about today. As we all know, at this moment in time, the world is being affected by the virus COVID-19, otherwise known as, the corona virus. Because of how contagious this virus is and how many people worldwide have been affected by it, there has been an increase in the number of patients in hospitals. This means that there would be a high demand in medical supplies and many of those hospitals are unable to receive the materials they need to be able to fight the virus. This is a problem as many of the doctors and hospital workers are risking their lives to save others. If they are unable to be fully shielded, they are in danger of contracting the virus themselves. My team and I have thought of an idea that would help make this problem have a solution. Our product will be an online medical store that is open to doctors to receive their medical supplies from so that the general public cannot stock on item and cause doctors to have a shortage. One specific problem that is being observed is that there is a shortage of N95 respirator masks, face shields and gowns. However, our plan would help prevent the general from buying out all the products especially during a time like this. However, in the future, we will make a few items available to the public but if there is a problem like this in the future, our store will be available only medical institutions